+++ src/js/api/serverless.mjs
@@ -38,7 +38,10 @@ import {
 
 import {
   setStoreValue,
-  removeStoreValue,
+  removeStoreValue
+} from '../lib/utils-html.mjs'
+
+import {
   epoch,
   anId
 } from '../lib/utils.mjs'